LPS-83808 Turn renderers into interfaces to minimize semver upgrades in the future
